backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
sqlalchemy==2.0.28
pyjwt==2.8.0
passlib==1.7.4
python-dotenv==1.0.1
bcrypt==4.1.2
alembic==1.13.1
psycopg2-binary==2.9.9
python-multipart==0.0.9
email-validator==2.1.0.post1
pydantic==2.6.3
pydantic-settings==2.2.1
